@@ -20,7 +20,7 @@ maintainers = [
     { name = "Timo Ege", email = "Timoege@online.de" },
 ]
 version = "0.0.1"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 license = { text = "Apache-2.0" }
 classifiers = [
     "Programming Language :: Python :: 3",
